Replace time.monotonic_ns with time.monotonic
This gets rid of the s_to_ns utility function, and more importantly,
allows the script to run on Python versions as old as 3.6.x (tested
with an Ubuntu Bionic container, 2018 vintage).

I originally used monotonic_ns as I wanted to avoid any precision
shenanigans when the numbers got very large, which is where floats
break down. Python integer numbers are represented as a type that
switches between a 64-bit long and a "bignum" type which can grow
as large as it wants to be with no loss of precision.

However, Python floating point numbers in practically all Python
implementations are mapped to double precision IEEE 754 double
precision floats, which (assuming they don't use the x86 extended
precision, unless the Wikipedia article implicitly assumes they
do) are accurate to integer precisions up to 2 to the power of 53,
which in terms of the timer is about 285616414 years.

That's probably a long enough system uptime to warrant a reboot to
reset the monotonic timer.